+++ src/main/java/com/researchspace/api/clientmodel/FolderTreeItemListing.java
@@ -3,13 +3,17 @@ package com.researchspace.api.clientmodel;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 
 @Data
 @EqualsAndHashCode(callSuper=true)
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class FolderTreeItemListing extends PaginatedResultList  {
 	
 	/**