+++ src/main/java/com/researchspace/api/clientmodel/UserSearchResult.java
@@ -7,16 +7,17 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 /**
- * FileSearchResult
+ * Paginated list of users returned by the sysadmin user-listing endpoint.
  */
 @Data
 @EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class FileSearchResult extends PaginatedResultList {
-
-    private List<ApiFile> files = new ArrayList<>();
+public class UserSearchResult extends PaginatedResultList {
+    private List<User> users = new ArrayList<>();
 
 }